refactor(app): moved ingest subcmds
